adding razorPay gateway

+++ src/main/java/com/lcwd/electronicStore/ElectronicStore/dtos/CartItemDto.java
@@ -12,6 +12,6 @@ public class CartItemDto {
     private int cartItemId;
     private ProductDto product;
     private int quantity;
-    private int totalPrice;
+    private long totalPrice;
 
 }
